Support for local business knowledge graph type and other organization subtypes.
- Adds address, image and geo fields to the knowledge graph object for
organizations. These are required by Google for LocalBusinesses but are
all valid fields in an Organization Object.
- Allows changing the organization "@type" to one of the LocalBusiness
subtypes

Note: where a file I was edited was mixed tabs and spaces, I have
sometimes changed that to just tabs as that seems to be the predominate
file. Appologies for the confusing diff due to that.